Reintroduce email format validation
We temporarily disabled it because of incompatibility between email_validation gem (the one we used to implement it) and email_validator gem (a clearance dependency we have just introduced with it). It turns out, both gems work the same way so we can just use the code we wrote with the new dependency.